+++ src/com/maddyhome/idea/vim/option/Options.java
@@ -33,6 +33,9 @@ import java.util.*;
  * Maintains the set of support options
  */
 public class Options {
+  public static final String RELATIVE_NUMBER = "relativenumber";
+  public static final String NUMBER = "number";
+
   /**
    * Gets the singleton instance of the options
    *
@@ -447,6 +450,8 @@ public class Options {
     addOption(new NumberOption("undolevels", "ul", 1000, -1, Integer.MAX_VALUE));
     addOption(new ToggleOption("visualbell", "vb", false));
     addOption(new ToggleOption("wrapscan", "ws", true));
+    addOption(new ToggleOption(NUMBER, "nu", false));
+    addOption(new ToggleOption(RELATIVE_NUMBER, "rnu", false));
   }
 
   private void addOption(@NotNull Option option) {