+++ CHANGES.md
@@ -7,6 +7,10 @@ History of changes in IdeaVim for the IntelliJ platform.
 0.38, TBD
 ---------
 
+Added support for `number` and `relativenumber` options.
+
+* VIM-410 Added support for 'relativenumber' option
+* VIM-483 Added support for 'number' option
 * VIM-624 Deselect visual selection range on opening the Ex entry field
 
 
@@ -14,7 +18,6 @@ History of changes in IdeaVim for the IntelliJ platform.
 ----------------
 
 A bugfix release.
-
 
 Bug fixes: